FINAL_PROJECT/FINAL_PROJECT/api/requirements.txt
Flask==2.2.5
boto3==1.28.0
python-docx==0.8.11
pdfplumber==0.7.7
werkzeug==2.2.3

